splitting was a bad idea because of multiple dependencies, returned back and moved to separate directory. added test controller and some services
