+++ quiz/util/dto.py
@@ -1,7 +1,7 @@
 import datetime
 from typing import List
 
-from miminet_model import Section, Test, User
+from quiz.entity.entity import Section, Test
 
 
 def to_section_dto_list(sections: List[Section]):
@@ -19,8 +19,7 @@ def to_test_dto_list(tests: List[Test]):
 
     for i in range(len(tests)):
         test = tests[i]
-        user = test.created_by_user()
-        dto_list.append(TestDto(test.name, test.created_by_user(), test.description))
+        dto_list.append(TestDto(test.id, test.name, test.created_by_user.email, test.description))
 
     return dto_list
 
@@ -33,7 +32,8 @@ class SectionDto:
 
 
 class TestDto:
-    def __init__(self, test_name: str, author_name: str, description: str) -> None:
+    def __init__(self, test_id: str, test_name: str, author_name: str, description: str) -> None:
+        self.test_id = test_id
         self.test_name = test_name
         self.author_name = author_name
         self.description = description
